requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ExifRead==3.0.0
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
